+++ config.js
@@ -65,18 +65,6 @@ require.config({
         "jquery"
       ]
     },
-    "amber-attic/IDE": {
-      "deps": [
-        "codemirror/lib/codemirror",
-        "codemirror/mode/smalltalk/smalltalk",
-        "codemirror/addon/hint/show-hint",
-        "css!codemirror/theme/ambiance",
-        "css!codemirror/lib/codemirror",
-        "css!codemirror/addon/hint/show-hint",
-        "jquery-ui",
-        "css!./resources/amber"
-      ]
-    },
     "ensure-console": {
       "exports": "console"
     }